requirements.txt
Django==2.2.8
psycopg2
Scrapy
djangorestframework
